feat: auth prep domain
